refactor(recall): verify/recall CLI を削除し計測を recall-bench crate へ分離 (ADR-0005)
yomu の product 責務は AI agent 向けコード検索。品質計測は maintainer 責務であり
公開 CLI surface は不要。両者を 1 つの CLI に混在させると責務が二重化する。verify/
recall を CLI から外し search-only にする (ADR-0004 の「CLI stays」サブ決定のみ amend、
Phase 5 cancellation / closure-scoping は据え置き)。

計測は決定性で分離:
- 決定的 (verify injection precision/recall、seeded recall) は既存 test gate
  (tests/verification.rs、tests/brief_recall.rs) に残す。
- 非決定的・実モデル必須の seed-less recall は workspace bench crate recall-bench へ。
  embedder は compile-time feature gate (test-support=stub / prod=real model) のため
  #[ignore] テストでは実モデル経路を組めず、yomu に依存する別 crate が必要。

- Cargo.toml: workspace 化 (members + default-members=["."]、resolver="3")。
  default-members により cargo install/release は yomu のみ配布、recall-bench は除外。
- crates/recall-bench: pub Yomu::recall を呼ぶ thin wrapper (可視性変更ゼロ)。
- main.rs: Command::Verify/Recall + dispatch + KNOWN_SUBCOMMANDS から除去。
- tools.rs: Yomu::verify_standalone 削除 (Yomu::recall / verify::measure は gate/bench が使用)。
- tests: recall_cli_* 4 件 + verify CLI 2 件削除。
- ci.yml: test/clippy/llvm-cov を -p yomu、check を --workspace に scope
  (recall-bench は test-support 未定義)。

nextest -p yomu --features test-support 713 pass。
