+++ Cargo.toml
@@ -1,3 +1,8 @@
+[workspace]
+members = [".", "crates/recall-bench"]
+default-members = ["."]
+resolver = "3"
+
 [package]
 name = "yomu"
 version = "0.16.0"
