+++ tests/cli_integration.rs
@@ -2057,82 +2057,3 @@ fn brief_json_emits_per_chunk_source_kind() {
         "FR-009a: at least one brief chunk must carry source_kind='src'",
     );
 }
-
-// ── verify subcommand (T-402, T-403) ────────────────────────────────
-
-// T-402: yomu --json verify emits structured precision/recall report
-// Spec FR-406b / FR-407.
-#[test]
-fn yomu_json_verify_emits_precision_recall_schema() {
-    let output = yomu_cmd()
-        .arg("--json")
-        .arg("verify")
-        .output()
-        .expect("yomu --json verify runs");
-    assert!(
-        output.status.success(),
-        "yomu --json verify must exit 0, status={:?} stderr={}",
-        output.status,
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let parsed: serde_json::Value =
-        serde_json::from_str(stdout.trim()).expect("verify JSON output parses");
-
-    for field in [
-        "precision",
-        "recall",
-        "tp",
-        "fp",
-        "fn_count",
-        "details",
-        "degraded",
-        "injection_check",
-    ] {
-        assert!(
-            parsed.get(field).is_some(),
-            "FR-406b: JSON output must include top-level field `{field}`, got: {parsed}"
-        );
-    }
-    assert_eq!(
-        parsed["injection_check"], "ran",
-        "FR-406b: injection_check must serialize to lowercase \"ran\", got: {parsed}"
-    );
-    // The bundled corpus must meet the gate at runtime as well as in the
-    // dedicated verification test.
-    let precision = parsed["precision"].as_f64().expect("precision is a number");
-    let recall = parsed["recall"].as_f64().expect("recall is a number");
-    assert!(
-        precision >= 0.90,
-        "BR-403 via CLI: precision must be >= 0.90, got {precision}: {parsed}"
-    );
-    assert!(
-        recall >= 0.95,
-        "BR-403 via CLI: recall must be >= 0.95, got {recall}: {parsed}"
-    );
-}
-
-// T-403: yomu verify plain text emits the header line and a per-entry table
-// Spec FR-406c / FR-407.
-#[test]
-fn yomu_verify_plain_text_emits_header_and_table() {
-    let output = yomu_cmd().arg("verify").output().expect("yomu verify runs");
-    assert!(
-        output.status.success(),
-        "yomu verify must exit 0, status={:?} stderr={}",
-        output.status,
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.starts_with("Verification: precision="),
-        "FR-406c: plain output must start with `Verification: precision=`, got: {stdout}"
-    );
-    // Per-entry table header line is the second meaningful line.
-    assert!(
-        stdout.contains("\nid\tkind\tmatched\texpected\tactual\n"),
-        "FR-406c: plain output must include the per-entry table header, got: {stdout}"
-    );
-}
